requirements1.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dateparser==1.0.0
dill==0.3.3
elmoformanylangs==0.0.4.post2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
joblib==0.17.0
Keras-Preprocessing==1.1.2
Markdown==3.3.3
noun-phrase-ua==0.0.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
overrides==3.1.0
pandas==1.1.4
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.7.1
torchvision==0.3.0
tzlocal==2.1
ufal.udpipe==1.2.0.3
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0